Clean repo from a lot of unecessary code as well as add comments to highscoreDaoTest
Fixed an error with highscore composite key

+++ Lab_3/lab_3/src/main/java/com/lab3/model/entity/key/CommentPK.java
@@ -22,7 +22,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class CommentPK implements Serializable{
     @Id
-    private long commentId;
+    private int commentId;
     @Id
     private String users;
     @Id
